requirements.txt
fastapi[all]
uvicorn[standard]
psycopg2
psycopg2-binary
sqlalchemy
bcrypt==4.0.1
passlib[bcrypt]
python-jose[cryptography]